Add a little bit of styling :)
