+++ src/components/app.css
@@ -1,0 +1,8 @@
+* {
+  padding: 0;
+  margin: 0;
+}
+
+.ui.header:first-child {
+  margin-top: 4em!important;
+}
